+++ PhpSerializerNET/PhpDeserializer.cs
@@ -19,6 +19,8 @@ namespace PhpSerializerNET {
 			{ "DateTime", typeof(PhpDateTime) }
 		};
 
+		private static Dictionary<Type, Dictionary<string, PropertyInfo>> PropertyInfoCache = new();
+
 		public PhpDeserializer(string input, PhpDeserializationOptions options) {
 			this._options = options;
 			if (this._options == null) {
@@ -34,6 +36,23 @@ namespace PhpSerializerNET {
 		public T Deserialize<T>() {
 			Type targetType = typeof(T);
 			return (T)this.DeserializeToken(targetType, _token);
+		}
+
+		/// <summary>
+		/// Reset the type lookup cache.
+		/// Can be useful for scenarios in which new types are loaded at runtime in between deserialization tasks.
+		/// </summary>
+		public static void ClearTypeCache() {
+			TypeLookupCache.Clear();
+			TypeLookupCache.Add("DateTime", typeof(PhpDateTime));
+		}
+
+		/// <summary>
+		/// Reset the property info cache.
+		/// Can be useful for scenarios in which new types are loaded at runtime in between deserialization tasks.
+		/// </summary>
+		public static void ClearPropertyInfoCache() {
+			PropertyInfoCache.Clear();
 		}
 
 		private object DeserializeToken(PhpSerializeToken token) {
@@ -76,7 +95,9 @@ namespace PhpSerializerNET {
 							break;
 						}
 					}
-					TypeLookupCache.Add(typeName, targetType);
+					if (_options.TypeCache.HasFlag(TypeCacheFlag.ClassNames)) {
+						TypeLookupCache.Add(typeName, targetType);
+					}
 				}
 			}
 			if (targetType != null && typeName != "stdClass") {
@@ -219,7 +240,16 @@ namespace PhpSerializerNET {
 
 		private object MakeObject(Type targetType, PhpSerializeToken token) {
 			var result = Activator.CreateInstance(targetType);
-			var properties = targetType.GetProperties().GetAllProperties(_options);
+			Dictionary<string, PropertyInfo> properties = null;
+
+			if (PropertyInfoCache.ContainsKey(targetType)) {
+				properties = PropertyInfoCache[targetType];
+			} else {
+				properties = targetType.GetProperties().GetAllProperties(_options);
+				if (_options.TypeCache.HasFlag(TypeCacheFlag.PropertyInfo)) {
+					PropertyInfoCache.Add(targetType, properties);
+				}
+			}
 
 			for (int i = 0; i < token.Children.Count; i += 2) {
 				var propertyName = _options.CaseSensitiveProperties ? token.Children[i].Value : token.Children[i].Value.ToLower();